chore: rename user-info folder to user_info
