+++ src/components/viewmodels/right_panel/user_info/admin/UserInfoKickButtonViewModel.tsx
