+++ test/unit-tests/components/viewmodels/right_panel/user_info/admin/UserInfoBanButtonViewModel-test.tsx
@@ -11,8 +11,8 @@ import { type Room, type MatrixClient, RoomMember } from "matrix-js-sdk/src/matr
 import { KnownMembership } from "matrix-js-sdk/src/types";
 
 import { MatrixClientPeg } from "../../../../../../../src/MatrixClientPeg";
-import { type RoomAdminToolsProps } from "../../../../../../../src/components/viewmodels/right_panel/user-info/admin/UserInfoAdminToolsContainerViewModel";
-import { useBanButtonViewModel } from "../../../../../../../src/components/viewmodels/right_panel/user-info/admin/UserInfoBanButtonViewModel";
+import { type RoomAdminToolsProps } from "../../../../../../../src/components/viewmodels/right_panel/user_info/admin/UserInfoAdminToolsContainerViewModel";
+import { useBanButtonViewModel } from "../../../../../../../src/components/viewmodels/right_panel/user_info/admin/UserInfoBanButtonViewModel";
 import Modal from "../../../../../../../src/Modal";
 import { withClientContextRenderOptions } from "../../../../../../test-utils";
